refactor: replace integer-semantic f64 fields with integer types across crates
Counts, durations, timestamps, and byte sizes that were modeled as f64 now
use u64/i64/u32/usize. f64 remains only where it is semantically a float:
ratios, percentages, scores, model params (temperature/top_p), monetary
values, and the JS-number-compat layers (cron, telemetry, legacy wire reads).

- protocol: add lenient integer deserializers (accept 100/100.0, truncate
  toward zero); fix range check so negative fractions like -0.1 are rejected
- agent-core-v2: goal budgets, task limits, turn/step durations, media crop
  regions, session/file timestamps, usage counters now integer-typed
- oauth/minidb/transcript: wire readers prefer as_u64/as_i64 with an f64
  fallback for legacy float spellings
- app: parse retryAfterMs as u64; file size parse keeps legacy "2.0" support

+++ apps/kimi-code/src-tauri/src/lib.rs
@@ -143,7 +143,7 @@ struct DeviceCodeEvent {
     user_code: String,
     verification_uri: String,
     verification_uri_complete: String,
-    expires_in: Option<f64>,
+    expires_in: Option<u64>,
 }
 
 impl From<DesktopDeviceCode> for DeviceCodeEvent {
